Refactor MoE directory
