+++ dataset/wikitext_perplexity.py
@@ -1,5 +1,5 @@
 import torch, math
-from mixture_of_experts.mixture_of_experts import DynamicMoE
+from MoE.mixture_of_experts import DynamicMoE
 from wikitext_loader import get_wikitext103
 from transformers import AutoTokenizer
